Remove keycloak-core and keycloak-crypto-default from SAML galleon feature pack and upgrade them to Java 17 closes #32586
