--- distribution/galleon-feature-packs/saml-adapter-galleon-pack/src/main/resources/modules/system/add-ons/keycloak/org/keycloak/keycloak-crypto-default/main/module.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-
-
-<!--
-  ~ Copyright 2022 Red Hat, Inc. and/or its affiliates
-  ~ and other contributors as indicated by the @author tags.
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~ http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<module xmlns="urn:jboss:module:1.3" name="org.keycloak.keycloak-crypto-default">
-    <resources>
-        <artifact name="${org.keycloak:keycloak-crypto-default}"/>
-    </resources>
-    <dependencies>
-        <module name="com.fasterxml.jackson.core.jackson-annotations"/>
-        <module name="com.fasterxml.jackson.core.jackson-core"/>
-        <module name="com.fasterxml.jackson.core.jackson-databind"/>
-        <module name="com.fasterxml.jackson.jaxrs.jackson-jaxrs-json-provider"/>
-        <module name="org.keycloak.keycloak-common" />
-        <module name="org.keycloak.keycloak-core" />
-        <module name="org.bouncycastle" />
-        <module name="org.jboss.logging"/>
-        <module name="javax.api"/>
-        <module name="jakarta.activation.api"/>
-        <module name="sun.jdk" optional="true" />
-    </dependencies>
-
-</module>
